Tested for new product

+++ src/main/java/pl/mprm/diet_calendar/service/ElementsService.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import pl.mprm.diet_calendar.dao.MacroElementRepository;
 import pl.mprm.diet_calendar.dao.MicroElementRepository;
+import pl.mprm.diet_calendar.dao.ProductRepository;
 import pl.mprm.diet_calendar.model.product_data.Product;
 
 @Service
@@ -13,10 +14,13 @@ public class ElementsService {
 
     private final MicroElementRepository microElementRepository;
     private final MacroElementRepository macroElementRepository;
+    private final ProductRepository productRepository;
 
     @Transactional
     public void deleteProductsElements(Product product) {
-        microElementRepository.deleteAllByProduct(product);
-        macroElementRepository.deleteAllByProduct(product);
+        if (product.getId() != null && productRepository.findById(product.getId()).isEmpty()) {
+            microElementRepository.deleteAllByProduct(product);
+            macroElementRepository.deleteAllByProduct(product);
+        }
     }
 }
